+++ desktopui/src/main/java/com/taim/desktopui/controllers/details/CustomerDetailsController.java
@@ -119,7 +119,7 @@ public class CustomerDetailsController implements IDetailController<CustomerDTO>
         this.stage = stage;
     }
 
-    public void initDataFromDB(int id){
+    public void initDataFromDB(long id){
         Task<List<TransactionDTO>> transactionTask = new Task<List<TransactionDTO>>() {
             @Override
             protected List<TransactionDTO> call() throws Exception {
