+++ desktopui/src/main/java/com/taim/desktopui/controllers/details/VendorDetailsController.java
@@ -108,7 +108,7 @@ public class VendorDetailsController implements IDetailController<VendorDTO> {
         this.stage = stage;
     }
 
-    public void initDataFromDB(int id){
+    public void initDataFromDB(long id){
         Task<List<TransactionDTO>> transactionTask = new Task<List<TransactionDTO>>() {
             @Override
             protected List<TransactionDTO> call() throws Exception {
